+++ CHANGES.rst
@@ -6,7 +6,8 @@ There's a frood who really knows where his towel is.
 1.0a2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add list of activities of a contact.
+  [hvelarde]
 
 
 1.0a1 (2014-10-02)